feature: StudyMemberCountService 추가
