+++ src/main/java/grep/neogulcoder/domain/study/service/StudyManagementServiceFacade.java
@@ -13,64 +13,17 @@ import org.springframework.transaction.annotation.Transactional;
 
 import static grep.neogulcoder.domain.study.exception.code.StudyErrorCode.*;
 
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class StudyManagementServiceFacade {
 
-    private final StudyRepository studyRepository;
-    private static final int MAX_RETRY = 3;
+    private final StudyMemberCountService studyMemberCountService;
 
-    @Transactional
     public void increaseMemberCount(Long studyId, Long userId) {
-        int retry = 0;
-        while (retry < MAX_RETRY) {
-            try {
-                Study study = getStudyById(studyId);
-                study.increaseMemberCount();
-                return;
-            } catch (OptimisticLockException | ObjectOptimisticLockingFailureException e) {
-                retry++;
-
-                try {
-                    Thread.sleep(30);
-                } catch (InterruptedException ex) {
-                    Thread.currentThread().interrupt();
-                    break;
-                }
-            }
-        }
-
-        log.warn("스터디 currentCount 증가 실패 (studyId={}, userId={})", studyId, userId);
-        throw new BusinessException(STUDY_MEMBER_COUNT_UPDATE_FAILED);
+        studyMemberCountService.increaseMemberCount(studyId, userId);
     }
 
-    @Transactional
     public void decreaseMemberCount(Long studyId, Long userId) {
-        int retry = 0;
-        while (retry < MAX_RETRY) {
-            try {
-                Study study = getStudyById(studyId);
-                study.decreaseMemberCount();
-                return;
-            } catch (OptimisticLockException | ObjectOptimisticLockingFailureException e) {
-                retry++;
-
-                try {
-                    Thread.sleep(30);
-                } catch (InterruptedException ex) {
-                    Thread.currentThread().interrupt();
-                    break;
-                }
-            }
-        }
-
-        log.warn("스터디 currentCount 감소 실패 (studyId={}, userId={})", studyId, userId);
-        throw new BusinessException(STUDY_MEMBER_COUNT_UPDATE_FAILED);
-    }
-
-    private Study getStudyById(Long studyId) {
-        return studyRepository.findById(studyId)
-            .orElseThrow(() -> new NotFoundException(STUDY_NOT_FOUND));
+        studyMemberCountService.decreaseMemberCount(studyId, userId);
     }
 }
